Add files via upload
These Scripts help to automate  the creation of zoom meetings based on your school timetable. We use Maze so the queries will have to be ammended to match your SMS. Our timetable data does not include start and end times of periods/blocks so we have to map block numers from the timetable data to times in the script. Clare Ciriaco Is responsible for creating these scripts.